// src/types/api/order.ts
export interface CustomerOrderSummary {
  id: number;
  created_at: string;
  updated_at: string;
  customer_name: string;
  customer_phone: string;
  order_number: string;
  order_status: number;
  order_status_string: string;
  payment_method: string;
  payment_status: string;
  sales_manager: string;
  shop: string;
  discount: number;
  due_amount: number;
  paid_amount: number;
  quantity: number;
  sub_total: number;
  total: number;
  consignment_id: string | null;
  tracking_code: string | null;
}